Review updates Part 1.
- Move downloadLogo and minifyLogoData into HIBP module.
- Fix logging in HIBP controller
- Lower acceptable response content-length max.
