+++ controllers/hibp.js
@@ -8,7 +8,6 @@ const EmailUtils = require("../email-utils");
 const HIBP = require("../hibp");
 const { LocaleUtils } = require ("../locale-utils");
 const HBSHelpers = require("../hbs-helpers");
-const BreachLogos = require("../breach-logos");
 
 const mozlog = require("../log");
 const log = mozlog("controllers.hibp");
@@ -40,12 +39,12 @@ async function notify (req, res) {
   // If we don't have a logo of the breach, get the logo.
   if (!fs.existsSync(`./public/img/logos/${breachAlert.LogoPath}`)) {
     try {
-      await BreachLogos.downloadLogo(breachAlert.LogoPath);
+      await HIBP.downloadLogo(breachAlert.LogoPath);
     } catch(err) {
       // Set LogoPath to missing-logo-icon.png if download fails
       // LogoPath will reset when the server restarts
-      breachAlert.LogoPath = "missing-logo-icon.png";
       log.error(`Unable to download logo for ${breachAlert.LogoPath}.`);
+      breachAlert.LogoPath = "missing-logo-icon.png";
     }
   }
